# specs/001-scope-van-de/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: ZzpChat MVP API
  description: API for ZzpChat WhatsApp AI Assistant for ZZP'ers
  version: 1.0.0
  contact:
    name: ZzpChat Support
    email: support@zzpchat.nl
servers:
  - url: https://api.zzpchat.nl/v1
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /invoices:
    get:
      summary: Get user invoices
      tags: [Invoices]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, SENT, PAID, OVERDUE, CANCELLED]
        - name: clientId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of invoices
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invoice'
                  total:
                    type: integer

    post:
      summary: Create new invoice
      tags: [Invoices]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, clientId]
              properties:
                amount:
                  type: number
                  minimum: 0.01
                clientId:
                  type: string
                description:
                  type: string
                dueDate:
                  type: string
                  format: date-time
                lineItems:
                  type: array
                  items:
                    type: object
                    properties:
                      description:
                        type: string
                      quantity:
                        type: number
                      price:
                        type: number
      responses:
        '201':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          description: Invalid input
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        company:
          type: string
        phone:
          type: string
        address:
          type: string
        vatNumber:
          type: string
        createdAt:
          type: string
          format: date-time
    Invoice:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [DRAFT, SENT, PAID, OVERDUE, CANCELLED]
        client:
          $ref: '#/components/schemas/Client'
        dueDate:
          type: string
          format: date-time
        sentAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
        description:
          type: string
        lineItems:
          type: array
          items:
            type: object
        createdAt:
          type: string
          format: date-time
...